Fix CompositeIterable
Make it compatible with Spliterator/Stream in java8 by returning a new
iterator each time iterator() is called
